+++ src/vscode/treeviews/Workspaces/FabricItemDefinitionFile.ts
@@ -55,8 +55,7 @@ export class FabricItemDefinitionFile extends FabricWorkspaceTreeItem {
 		}
 	}
 
-	public async delete(confirmation: "yesNo" | "name" | undefined = undefined): Promise<void> {
-		// currently disabled - TreeView is read-only!
-		this.definitionRoot.delete(confirmation, this);
+	get canDelete(): boolean {
+		return false;
 	}
 }